mosaic-maker/requirements.txt
flask
click
flask-restplus
Pillow
numpy
flask-cors
requests_futures
gunicorn
git+git://github.com/rjvanvoorhis/photomosaic.git